sync somehow works

+++ tasks/lab_1/common.h
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 
-void rectangle_step(double **field, uint64_t x_i, uint64_t t_i);
-void rectangle_scheme(double **field, uint64_t t_0, uint64_t t_n, uint64_t x_0, uint64_t x_n);
+void rectangle_step(double **field, int32_t x_i, int32_t t_i);
+void rectangle_scheme(double **field, int32_t t_0, int32_t t_n, int32_t x_0, int32_t x_n);
 
 #endif
